+++ sdk/keyvault/azure-security-keyvault-keys/src/delete_key_operation.cpp
@@ -27,12 +27,13 @@ inline Azure::Core::OperationStatus CheckCompleted(Azure::Core::Http::RawRespons
 }
 } // namespace
 
-Azure::Core::Http::RawResponse const&
+std::unique_ptr<Azure::Core::Http::RawResponse>
 Azure::Security::KeyVault::Keys::DeleteKeyOperation::PollInternal(Azure::Core::Context& context)
 {
+  std::unique_ptr<Azure::Core::Http::RawResponse> rawResponse;
   if (!IsDone())
   {
-    m_rawResponse = m_pipeline->GetResponse(
+    rawResponse = m_pipeline->GetResponse(
         context, Azure::Core::Http::HttpMethod::Get, {_detail::DeletedKeysPath, m_value.Name()});
     m_status = CheckCompleted(*m_rawResponse);
   }
@@ -40,7 +41,7 @@ Azure::Security::KeyVault::Keys::DeleteKeyOperation::PollInternal(Azure::Core::C
   // To ensure the success of calling Poll multiple times, even after operation is completed, a
   // copy of the raw http response is returned instead of transfering the ownership of the raw
   // response inside the Operation.
-  return *m_rawResponse;
+  return rawResponse;
 }
 
 Azure::Security::KeyVault::Keys::DeleteKeyOperation::DeleteKeyOperation(