+++ sdk/storage/azure-storage-files-shares/inc/azure/storage/files/shares/share_responses.hpp
@@ -220,13 +220,13 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
       std::abort();
     }
 
-    Azure::Core::Http::RawResponse const& PollInternal(Azure::Core::Context& context) override;
+    std::unique_ptr<Azure::Core::Http::RawResponse> PollInternal(
+        Azure::Core::Context& context) override;
 
     Azure::Response<Models::GetShareFilePropertiesResult> PollUntilDoneInternal(
         std::chrono::milliseconds period,
         Azure::Core::Context& context) override;
 
-    std::unique_ptr<Azure::Core::Http::RawResponse> m_rawResponse;
     std::shared_ptr<ShareFileClient> m_fileClient;
     Models::GetShareFilePropertiesResult m_pollResult;